Revert "chore(calculator): rename Level -> LevelRecords"
This reverts commit f5cec849d1014c2864358ba7a38cd79c98b3d71e.
