+++ app/src/Calculator/dashboard-frames/calculatorPageDashboardContents.ts
@@ -1,5 +1,5 @@
 import { DashboardItemProps } from '@shared/types/Dashboard';
-import { Blackhole, LevelRecords } from '../dashboard-contents';
+import { Blackhole, Level } from '../dashboard-contents';
 
 export const calculatorPageDashboardContents: DashboardItemProps[] = [
   {
@@ -8,6 +8,6 @@ export const calculatorPageDashboardContents: DashboardItemProps[] = [
   },
   {
     id: 1,
-    content: LevelRecords,
+    content: Level,
   },
 ];
